engine/requirements.txt
# TaxPulse Engine Dependencies

# Core dependencies
PyYAML>=6.0.1  # YAML parsing for rules and configs
json-logic-py>=1.5.0  # JSONLogic evaluation (optional - we have custom impl)

# Data processing
pandas>=2.1.0  # Transaction data handling
numpy>=1.24.0  # Numerical computations

# Development dependencies
pytest>=7.4.0  # Unit testing
pytest-cov>=4.1.0  # Test coverage
black>=23.12.0  # Code formatting
flake8>=6.1.0  # Linting
mypy>=1.7.0  # Type checking
